add sign-in page to stage page
