+++ components/sign-in-button.module.css
@@ -1,0 +1,31 @@
+.signin {
+  width: 80%;
+  height: 56px;
+  margin-top: var(--space-4x);
+  border-radius: var(--space-2x);
+  border: 1px solid var(--brand);
+  background: var(--brand);
+  cursor: pointer;
+  font-family: inherit;
+  font-size: var(--text-md);
+  letter-spacing: -0.02em;
+  outline: none;
+  font-weight: 500;
+  color: white;
+  transition: background-color 0.2s ease, color 0.2s ease;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.signin:hover,
+.signin:focus {
+  background: black;
+  color: var(--brand);
+}
+
+@media (min-width: 600px) {
+  .signin {
+    width: 150px;
+  }
+}
